+++ app/Http/Controllers/Api/UserOperationController.php
@@ -2,26 +2,25 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Exports\CollectionExport;
-use App\Factories\UserOperationStrategyFactory;
-use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreUserOperationRequest;
-use App\Http\Resources\UserOperationResource;
-use App\Models\Attachment;
-use App\Models\BlackList;
-use App\Models\Setting;
-use App\Models\UserOperation;
-use App\Services\Search;
 use Carbon\Carbon;
-use Illuminate\Http\Request;
+use App\Models\Setting;
+use App\Services\Search;
+use App\Models\BlackList;
+use App\Models\Attachment;
 use Illuminate\Support\Arr;
+use Illuminate\Http\Request;
+use App\Models\UserOperation;
+use App\Exports\CollectionExport;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\URL;
-use Intervention\Image\Facades\Image;
+use App\Http\Controllers\Controller;
 use Maatwebsite\Excel\Facades\Excel;
+use Intervention\Image\Facades\Image;
+use Illuminate\Support\Facades\Storage;
+use App\Http\Resources\UserOperationResource;
+use App\Factories\UserOperationStrategyFactory;
 use Illuminate\Pagination\LengthAwarePaginator;
-use Illuminate\Support\Collection;
+use App\Http\Requests\StoreUserOperationRequest;
 
 class UserOperationController extends Controller
 {
@@ -29,7 +28,7 @@ class UserOperationController extends Controller
 
     public function __construct()
     {
-        $this->authorizeResource(UserOperation::class);
+        // $this->authorizeResource(UserOperation::class);
         $this->search = new Search();
     }
 
@@ -51,10 +50,28 @@ class UserOperationController extends Controller
 
     public function index(Request $request, $id = null)
     {
+        $this->authorize('viewAny', UserOperation::class);
+
         $type = $request->input('type');
         $strategy = UserOperationStrategyFactory::createStrategy($type, $id);
 
         $data = $strategy->getUserOperations();
+
+        // Get the pagination limit from the request, default to 10 if not provided
+        $perPage = 200;
+
+        // Manually create a LengthAwarePaginator instance
+        $currentPage = LengthAwarePaginator::resolveCurrentPage();
+        $slicedData = $data->slice(($currentPage - 1) * $perPage, $perPage);
+        $paginator = new LengthAwarePaginator($slicedData, $data->count(), $perPage, $currentPage);
+
+        // return $paginator;
+        return [
+            $paginator->items(),
+            ['previousPageUrl' => $paginator->previousPageUrl(),
+            'nextPageUrl' => $paginator->nextPageUrl(),
+            'totalPages' => $paginator->lastPage(),]
+        ];
 
         // Get the pagination limit from the request, default to 10 if not provided
         $perPage = 200;
@@ -83,6 +100,8 @@ class UserOperationController extends Controller
 
     public function store(StoreUserOperationRequest $request)
     {
+        $this->authorize('create', UserOperation::class);
+
         return DB::transaction(function () use ($request) {
             $userOperation = UserOperation::create(Arr::except($request->validated(), ['wallet_photo']));
 
@@ -103,6 +122,8 @@ class UserOperationController extends Controller
      */
     public function checkUserForSanction(mixed $userOperation, $request)
     {
+        $this->authorize('viewAny', UserOperation::class);
+
         $settings = Setting::select('usd_to_som as usd', 'usdt_to_som as usdt', 'rub_to_som as rub', 'limit')->first()->toArray();
         $currencyRate = 1;
         if ($request->has('currency')) {
@@ -141,6 +162,8 @@ class UserOperationController extends Controller
 
     public function attach($photos, $user, $type)
     {
+        $this->authorize('create', UserOperation::class);
+
         $thumbnail_url = null;
         foreach ($photos as $key => $singleFile) {
 
@@ -165,6 +188,8 @@ class UserOperationController extends Controller
 
     public function range(Request $request)
     {
+        $this->authorize('create', UserOperation::class);
+
         $date1 = Carbon::createFromFormat('d/m/Y H:i', $request->date1)->format('Y-m-d H:i');
         $date2 = Carbon::createFromFormat('d/m/Y H:i', $request->date2)->format('Y-m-d H:i');
         $time1 = Carbon::createFromFormat('d/m/Y H:i', $request->date1)->format('Y-m-d');
@@ -184,6 +209,8 @@ class UserOperationController extends Controller
      */
     public function show(UserOperation $userOperation)
     {
+        $this->authorize('viewAny', $userOperation);
+
         return new UserOperationResource($userOperation->loadMissing(['addedUser', 'legalEntity']));
     }
 
@@ -196,6 +223,8 @@ class UserOperationController extends Controller
      */
     public function update(StoreUserOperationRequest $request, UserOperation $userOperation)
     {
+        $this->authorize('update', $userOperation);
+
         $userOperation->update($request->validated());
 
         return new UserOperationResource($userOperation);
@@ -209,12 +238,16 @@ class UserOperationController extends Controller
      */
     public function destroy(UserOperation $userOperation)
     {
+        $this->authorize('delete', $userOperation);
+
         $userOperation->delete();
         return response()->noContent();
     }
 
     public function search(Request $request)
     {
+        $this->authorize('viewAny', UserOperation::class);
+
         return  $userOperation = UserOperation::with(['addedUser', 'legalEntity'])
             ->when($request->get('type') == 'user' || $request->get('type') == null, function ($q) use ($request) {
                 $addedUsers = $this->search->searchFromClients('AddedUser', $request)->pluck('id');
